Extract assembly model to separate data from logic
- Created `src/semantic/assembly_model.rs` and moved `AssembledStatement`, `Constituent`, `VerbConstituent`, `ParticipleConstituent`, and `Literal` there.
- Updated `src/semantic/assembler.rs` to import these types instead of defining them.
- Updated `src/semantic/mod.rs` to expose the new module and update re-exports.
- Removed unused imports in `src/semantic/assembler.rs`.
- Fixed a duplicate import in `src/parser.rs`.

This refactor adheres to the "Atlas" persona by reducing the "Blob" anti-pattern in `assembler.rs` and separating data models from assembly logic.
